requirements.txt
asgiref==3.8.1
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
Unipath==1.1
uritemplate==4.1.1
